Convert interface in action to float64 first

+++ actions.go
@@ -57,7 +57,7 @@ func mockAction(i interface{}) {
 }
 
 func setStoveTemp(i interface{}) {
-	Stove.SetTemp(i.(int))
+	Stove.SetTemp(int(i.(float64)))
 }
 
 // Actions for the trigger-queue to execute
